refactor(cli): extract plugin installer
